feat(repository): create currency repository interface
This commit adds the initial Currency Repository interface with the
following methods:

- GetByCode: will receive the currency code and return its value or an
error.
- Create: will receive the currency data and create the currency entry
and will return either nil or an error.
- Update: will receive the currency data and update a currency and will
return either nil or an error.
- Delete: will receive the currency code delete a currency entry and
will return either nil or an error.

This is an initial layout of the currency repository, it could change
during the development proccess